refactor: update imports to use centralized _pydantic module for improved consistency
